requirements.txt
# Core AI/ML
openai>=1.0.0
anthropic>=0.8.0
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.3.0

# RAG & Vector Database
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
pypdf>=3.17.0
pymupdf>=1.23.0

# Multimodal Processing
pillow>=10.1.0
pytesseract>=0.3.10
pdf2image>=1.16.3
opencv-python>=4.9.0

# Fine-tuning
datasets>=2.16.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
trl>=0.7.0

# Web Framework
streamlit>=1.29.0
gradio>=4.13.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0